+++ ChangeLog.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Version [0.15.0](https://github.com/cedx/base/compare/v0.14.1...v0.15.0)
+- Added the `Open` property to the `UI.Components.Toast` component.
 - Renamed the `start()` and `stop()` methods of the `UI.Components.LoadingIndicator` component to `show()` and `hide()`.
 
 ## Version [0.14.1](https://github.com/cedx/base/compare/v0.14.0...v0.14.1)